requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-dotenv
docker
requests
passlib[bcrypt]
python-jose[cryptography]
pydantic[email]
pydantic-settings
python-multipart
cryptography
httpx